feat: Created backgroundColours utility & used it in ProductList & VideoList components
