+++ src/frontend/src/utils/backgroundColours.js
@@ -1,0 +1,11 @@
+// List of different background colours for each product, each product or video gets a different one based on its index; used in ProductList & VideoList components.
+const backgroundColours = [
+    "bg-[#EDF7FF]/70",
+    "bg-[#F4E6E6]/60",
+    "bg-[#F3F5DD]/70",
+    "bg-[#E3F1ED]/70",
+    "bg-[#F5F5F5]/70",
+    "bg-[#FFDEC4]/40"
+];
+
+export default backgroundColours;
